Refactor and fix profiling codes
