+++ docling/datamodel/document.py
@@ -3,10 +3,9 @@ import re
 from enum import Enum
 from io import BytesIO
 from pathlib import Path, PurePath
-from typing import TYPE_CHECKING, Dict, Iterable, List, Optional, Tuple, Type, Union
+from typing import TYPE_CHECKING, Dict, Iterable, List, Optional, Type, Union
 
 import filetype
-import numpy as np
 from docling_core.types.doc import (
     DocItem,
     DocItemLabel,
@@ -53,6 +52,7 @@ from docling.datamodel.base_models import (
     Page,
 )
 from docling.datamodel.settings import DocumentLimits
+from docling.utils.profiling import ProfilingItem
 from docling.utils.utils import create_file_hash, create_hash
 
 if TYPE_CHECKING:
@@ -178,29 +178,6 @@ class InputDocument(BaseModel):
 class DocumentFormat(str, Enum):
     V2 = "v2"
     V1 = "v1"
-
-
-class ProfilingScope(str, Enum):
-    PAGE = "page"
-    DOCUMENT = "document"
-
-
-class ProfilingItem(BaseModel):
-    scope: ProfilingScope
-    count: int = 0
-    times: List[float] = []
-
-    def avg(self) -> float:
-        return np.average(self.times)  # type: ignore
-
-    def std(self) -> float:
-        return np.std(self.times)  # type: ignore
-
-    def mean(self) -> float:
-        return np.mean(self.times)  # type: ignore
-
-    def percentile(self, perc: float) -> float:
-        return np.percentile(self.times, perc)  # type: ignore
 
 
 class ConversionResult(BaseModel):
